began to tweak the mouse enter & leave code while debugging the menubar
in TWindow::createWindow() we create a tracking window for all windows.  as
far as i remember, i've added that for the enter/leave messages.  i've now
noticed, that a mouseEnter message isn't just send for the deepest window,
but also for it's parents.  this is different from the behaviour under X11.

i might want to avoid the tracking window and traverse the window tree on
our own as this might be more efficient.