fixed the menubar's 1st click closes popup again, 2nd click is okay
did this by more fixes to the mouse enter|leave code.  it also doesn't work,
when the menubar doesn't have the focus and just receives it when getting
the focus.

i assume this is because in that case, the application doesn't get any mouse
events till the moment it receives the focus.

in that situation we need to send a mouse enter message without a mouse
button flag.