kludge for the help menu
the help menu became disabled after clicking once on the toolbar, and once
on the window with the menubar.  this was because it help menu entires had
an activation of PARENT_FOCUS. we now switch automatically to DOMAIN_FOCUS
when the action's parent has no parent itself.

this needs some more brain thrown upon later