TFConnection: began support for more than two points
o updatePoints() and bounds() still need to be adjusted
o also a way to delete points (ie. mark and hit delete or move point over another point)

while i'm astonished that i'm able to implement this with the old design of
TFigureEditor, i really have to redesign the whole figure editor and it's
associated classes.

and i will draw the new design with this editor. one of the reasons i am
continuing with this mess for a while.