toolbar: do not point to unreferenced objects
this is far from being final. and the code looks nasty and the TSlot idea
doesn't look so nice anymore.

the current idea on how to proceed is this:

the figure tools will continue to be singletons, hence the GChoice object
won't be bound to TTestToolbar.  instead it will be a global object.

i will then create two instances of TTestToolbar to use the globale GChoice
object.

then two instances of TTestToolbar with different sets of tools. the last
instance which received the focus will define which tools are available in
the toolbar.  (do we really need that?  use case is to run TLayoutEditDialog
and TFischland at the same time, or a computer network map editor and an
icon editor for the icons within the map.  but then those would also have
distinctive toolbars.)

but how to handle a selected tool, which is not available in TTestToolbar
which just received the focus?

another feature i wanted to add was a button to lock a window to a certain
figure tool, ie.  one drawing locked to pick figure attributes.

that feature would add a pointer to a figure tool, which could also be used
as a 'last valid figure tool' reference.